docs: Add comprehensive CSS and TypeScript migration guides
Added detailed documentation for frontend development best practices and TypeScript migration strategies.

New Documentation Files:

- CSS_GUIDELINES.md: Complete CSS Module best practices guide
  * File naming conventions (ComponentName.module.css)
  * Directory structure recommendations
  * camelCase class naming standard
  * CSS organization patterns
  * Decision trees for CSS placement
  * Comparison with other frameworks (Tailwind, Material UI, shadcn)
  * Migration path options for future
  * Quick reference tables and checklists

- TYPESCRIPT_MIGRATION_PLAN.md: Step-by-step TypeScript migration guide
  * 4-phase migration strategy (Setup → Services → Components → Testing)
  * Complete type definitions for all API responses, User, Auth, Banner
  * Before/after code comparisons (JavaScript vs TypeScript)
  * Service migration examples (apiBaseUrlProvider, refreshTokenService, axiosRequestHandler)
  * Component migration examples (Banner, Login, Register, Dashboard)
  * Common type patterns (Props, useState, Event Handlers, API responses)
  * Troubleshooting section with solutions
  * Full migration checklist
  * Post-migration best practices

- CLEAN_TYPESCRIPT_SETUP.md: Fresh TypeScript project setup guide
  * 12-phase clean migration plan
  * Fresh Vite + TypeScript project creation
  * Complete folder structure setup
  * TypeScript configuration (tsconfig.json, tsconfig.app.json)
  * All type definitions with proper interfaces
  * Full code examples for all services and components
  * Step-by-step copy and convert strategy
  * Git workflow for clean migration
  * Testing and validation procedures
  * Troubleshooting common issues
  * Complete migration checklist

Benefits:

- Provides team with clear CSS architecture standards
- Offers two migration paths (gradual vs fresh) for TypeScript adoption
- Includes complete code examples for all components and services
- Establishes best practices for growing codebase
- Improves developer onboarding with comprehensive documentation
- Sets foundation for scalable, maintainable frontend architecture
