+++ Backend/.env.example
@@ -1,0 +1,3 @@
+# MDSystem Backend Environment Variables
+# Copy this file to .env and fill in your values
+# .env is git-ignored and will not be committed to the repository
