+++ mdsystem-frontend/.env.example
@@ -1,0 +1,25 @@
+# MDSystem Frontend Environment Variables
+# Copy this file to .env and fill in your values
+# .env is git-ignored and will not be committed to the repository
+
+# ============================================
+# Local Development
+# ============================================
+# Backend API URL for local development
+VITE_API_URL=http://localhost:3001
+
+# ============================================
+# Production URLs (mdsystemtip.space)
+# ============================================
+# Patient Portal API (www subdomain)
+VITE_PATIENT_API_URL=https://www.mdsystemtip.space
+
+# Staff Portal API (staff subdomain)
+VITE_STAFF_API_URL=https://staff.mdsystemtip.space
+
+# ============================================
+# Notes:
+# ============================================
+# - All variables must start with VITE_ to be accessible in the app
+# - These are used by src/services/apiBaseUrlProvider.js
+# - If not set, the app will use dynamic hostname detection as fallback
